chore: refactor plugin data fetch to reduce duplication
